backend/requirements.txt
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Markdown==3.2.1
Pillow==7.1.2
PyJWT==1.7.1
pytz==2019.3
wrapt==1.11.2
djoser
drfaddons>=0.1.0
requests>=2.21.0